# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Tracker API
  version: 1.0.0
  description: JSON REST API for the Medical Tracker Application MVP (api_contract.md).
paths:
  /api/v1/account/:
    get:
      operationId: account_retrieve
      description: |-
        `GET` / `PATCH /api/v1/account/` — always the authenticated user's own account.

        There is no identifier in the path, so no other account can be addressed
        (ADS-FR-009-01). Changing the timezone leaves stored instants untouched.
      tags:
      - account
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
    patch:
      operationId: account_partial_update
      description: |-
        `GET` / `PATCH /api/v1/account/` — always the authenticated user's own account.

        There is no identifier in the path, so no other account can be addressed
        (ADS-FR-009-01). Changing the timezone leaves stored instants untouched.
      tags:
      - account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAccountUpdateRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
          description: 'Validation failure (api_contract.md §3.4): each field name,
            or `non_field_errors`, maps to its messages.'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Authentication failure (api_contract.md §3.1).
...
components:
  schemas:
...
    Account:
      type: object
      description: 'The account representation: `id`, `email`, `timezone` (api_contract.md
        §6.1).'
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        timezone:
          type: string
          readOnly: true
      required:
      - email
      - id
      - timezone
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
...
    PatchedAccountUpdateRequest:
      type: object
      description: '`PATCH /api/v1/account/`: only `timezone` is writable (api_contract.md
        §6.2).'
      properties:
        timezone:
          type: string
          minLength: 1
          maxLength: 64
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT